Implement interfaces for eudev

+++ libc/generic/posix/include/mlibc/dev_t.h
@@ -2,7 +2,7 @@
 #ifndef MLIBC_DEV_T_H
 #define MLIBC_DEV_T_H
 
-typedef int dev_t;
+typedef unsigned long dev_t;
 
 #endif // MLIBC_DEV_T_H
